Removed Rule that Broke the Website
The CSS rule "overflow-x: hidden" in the body tag was creating issues across several pages with the Hero Section elements remaining on screen past their container locations. This rule was added to maintain the visual affect  of the h1 and h4 backgrounds on the Portfolio Page. Solution was to move the offended rule to the element specific container, rather than to apply to the entire body element.