Connects Pages to About Me and Adds Content
Adds the About Me option to the navigations on all pages to get to the About Me Page. Created a button on the home page in the "Introduction" section that leads to the About Me Page. And added the top content for the About Me page; a small summary of who I am and a thematic quote for the top.