Fixes Hero Section Error
The portfolio page Hero Section was only displaying with a height of 35rem, hardcoded into the html. This wasn't a problem before, but seems to have made itself present upon setting the body to a display flex rule. Either way, it has been corrected to display a height of 100vh.